int!: remove sdk.v2 module (#305)
# The problem

We had a bunch of docs builds warnings related to `orquestra.sdk.v2`.

# This PR's solution

Removes `orquestra.sdk.v2`. It's been a while since we needed it, and
this module isn't kept up-to-date with the public SDK API anyway.

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/imp[rovement]/int[ernal]/docs>[!]:`
- [x] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).