feat!:  add orquestra sdk head node current version as a requirement in ray runtime environments (#324)
# The problem

When a workflow runs on CE/Ray, there is a head node and worker nodes.
The head node will use the version of orquestra-sdk in the environment
that launched Ray (or `orq up`). Each worker can install a new python
environment per task.

It’s possible for the dependencies of a task to install the wrong
version of orquestra-sdk inside a worker, e.g. by a wrong custom image
or via poorly defined PythonImports.

The outcome of ORQSDK-848 was to add orquestra-sdk==<head node
version>to the runtime environment of all tasks.

If a different version is brought in via the user’s imports, pip will
complain and will quit. See the ADR in 848 for more details.

# This PR's solution

Adds the currently installed[^1] SDK version to the runtime environment
pip requirements.

[^1]: In order to avoid trying to install
`orquestra-sdk==0.57.1.dev3+g3ef9f57.d20231003`, we strip out prerelease
versioning (in this example, the correct dependency
`orquestra-sdk==0.57.0` would be added instead). This creates an issue
in our CI where the installed SDK version detected by setuptools is
`0.1-dev...`. I've therefore implemented an escape hatch where rather
than the semantically nonsensical `orquestra-sdk==0.0`, this case adds
`orquestra-sdk` without a version. This somewhat compromises the intent
of this PR, but I can't think of a scenario outside our CI where this
scenario would occur.

As a result of this change, specifying the SDK in the task dependencies
is now meaningless - it will always be installed regardless of what you
do. Therefore, before adding the current version to the list we check
for a dependency on the SDK and, if it's there, remove it from the list
and warn the user that it has been overruled. This means that rather
than getting a pip error when listing the SDK as a requirement, the user
gets an informative message and the workflow runs as intended. This is
intended to reduce the confusion of pip telling them that there are
multiple clashing requirements when they've only specified one.

I've also updated the `dependency installation` guide so it no longer
indicates that people can specify sdk versions in their dependencies.



# Supplemental

I've also had to increase the polling timeout for integration tests -
all tasks now require _at minimum_ the SDK to install, so they take a
little longer.

# Checklist

_Check that this PR satisfies the following items:_

- [X] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [X] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [X] The PR's title is prefixed with
`<feat/fix/chore/imp[rovement]/int[ernal]/docs>[!]:`
- [X] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).

[ORQSDK-937]

[ORQSDK-937]:
https://zapatacomputing.atlassian.net/browse/ORQSDK-937?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: James Clark <70290797+jamesclark-Zapata@users.noreply.github.com>