feat: find config by querying runtimes (#332)
Split from
https://github.com/zapatacomputing/orquestra-workflow-sdk/pull/328.

# The problem

We want to eventually get rid of client-side workflow runs SQLite DB. We
had a research spike few weeks ago. One of action items was to change
the rules for resolving config name if the user didn't pass it
explicitly.

# This PR's solution

If the user doesn't pass `config` directly, we ask all known runtimes
for this `wf_run_id`. If no runtime says "yeah, I know this run", we
present an error to the user. The error message contains a description
of which runtimes have been queried. I don't consider this a breaking
change. Affected APIs:
* `sdk.WorkflowRun.by_id(wf_run_id="...")`
* `orq wf view <wf_run_id>`

As discussed in
https://github.com/zapatacomputing/orquestra-workflow-sdk/pull/328, I
first implemented the functionality change as requested in the JIRA
ticket. When this PR is merged, I'm planning to follow up with a
refactor to make the tests more maintainable.

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/imp[rovement]/int[ernal]/docs>[!]:`
- [x] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).

---------

Co-authored-by: Sebastian Morawiec <Sebastian.Morawiec@zapatacomputing.com>