Int: Remove optional str from runtime name (#340)
# The problem

https://zapatacomputing.atlassian.net/browse/ORQSDK-996
Name for config was optional for historical reasons. Now we always pass
the name in our factory methods.

# This PR's solution

Initial idea was to deprecate usage of `name = None` BUT
I found out that our `RuntimeConfig` has convention of factory methods,
and using `__init__` should not be considered as public API due to
`bypass_factory_methods` setting (plus there is bunch of comments saying
to not use it)
So as a solution ive already changed the `name` parameter from
`optional[str]` to `str`

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/imp[rovement]/int[ernal]/docs>[!]:`
- [x] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).