int: declutter PR check view (#350)
# The problem

Our CI runs so many jobs that the PR checks list is getting difficult to
browse. Meanwhile, not every combination of job-OS-Python brings benefit
of running on every push, for every PR.

# This PR's solution

* Remove Python 3.10 from the PR set up matrix. We'll run against Python
3.11 and 3.9.
* Bump performance tests workflow to Python 3.11.

I'll remove the 3.10 job requirement from the repo settings when this PR
is accepted.

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/imp[rovement]/int[ernal]/docs>[!]:`
- [x] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).