Feat: Change statuses reporting (#375)
# The problem
We were overwriting Ray FAILED status with RUNNING in some cases which
lead to WDR issues.
https://zapatacomputing.atlassian.net/browse/ORQSDK-1024

# This PR's solution
Change how we report status of a workflow where 1 task already failed,
but there are still tasks running. It uses to be RUNNING now we go back
to FAILED

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/imp[rovement]/int[ernal]/docs>[!]:`
- [x] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).