fix: log dumpers (#383)
# The problem

Dumping logs in the CLI no longer worked after we changed logging
behaviour a while ago.

Note: this was spotted by Pyright! But it was disabled.

# This PR's solution

Updates the log dumper to work with out + err logs.

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/imp[rovement]/int[ernal]/docs>[!]:`
- [x] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).