feat: add cli command to generate graphiz graph of a workflow (#404)
# The problem

_Please describe the motivation for this PR. Bonus points for code
examples, error messages, and screenshots._

# This PR's solution

Adds the following CLI command:

```bash
orq wf graph 
```

There are two potential use cases - either we want to display the graph
for a local workflow definition, or we want to display it for a
previously submitted workflow. These are handled in the following ways:

## Case 1: Local workflow definition

This can be specified with at least one of `MODULE` and `NAME`:

```bash
orq wf graph [-m MODULE ] [-n NAME]
```

If the name or module are omitted, the user will be prompted to enter
it.

## Case 2: Previously submitted workflow

Any combination of the workflow run id, config, and workspace can be
specified. The CLI will prompt the user to select the missing
information in the same way as `orq wf status`.

```bash
orq wf graph [-c CONFIG] [-w WORKSPACE] [--id WORKFLOW_RUN_ID]
```

finally, in parallel with `orq wf status`, the command

```bash
orq wf graph
```

with no arguments prompts the user to select a previously submitted
workflow run to display.

## Specifying the output file

The output file can be specified using [-f FILEPATH] in any case.

# Checklist

_Check that this PR satisfies the following items:_

- [x] Tests have been added for new features/changed behavior (if no new
features have been added, check the box).
- [x] The [changelog file](CHANGELOG.md) has been updated with a
user-readable description of the changes (if the change isn't visible to
the user in any way, check the box).
- [x] The PR's title is prefixed with
`<feat/fix/chore/imp[rovement]/int[ernal]/docs>[!]:`
- [x] The PR is linked to a JIRA ticket (if there's no suitable ticket,
check the box).

[ORQSDK-1008]


[ORQSDK-1008]:
https://zapatacomputing.atlassian.net/browse/ORQSDK-1008?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ